requirements.txt
Flask==2.3.2
flask[async]==2.3.2
uvicorn==0.23.2
pymongo==4.5.0
Authlib==1.2.1
cloudinary==1.36.0
requests==2.31.0
python-dotenv==1.0.0